everything but streaming block test
