--- tempodb/encoding/v0test/fake/00f5a116-639e-4880-bbe7-be9b0c828033/meta.json
@@ -1,1 +1,0 @@
-{"format":"v0","blockID":"00f5a116-639e-4880-bbe7-be9b0c828033","minID":"EG8kdVDmE2np+sFQ1Ai9xA==","maxID":"83bQxf9aEu5hyf6dWl+suA==","tenantID":"fake","startTime":"2021-01-28T15:20:33.579379225-05:00","endTime":"2021-01-28T15:30:33.579379554-05:00","totalObjects":10,"compactionLevel":0}
