--- tempodb/encoding/v0test/fake/00f5a116-639e-4880-bbe7-be9b0c828033/readme.md
@@ -1,5 +1,0 @@
-# v0 block
-
-This block was created manually using the original v0 block code.  It is vaguely expected for v0 code to drift over time
-as bug fixes are introduced, changes are made for v1, or the backend code changes.  This block and the associated tests
-will increase confidence that these changes do not impact our ability to read a v0 block.
