--- tempodb/encoding/v1/record.go
@@ -1,10 +1,0 @@
-package v1
-
-import (
-	"github.com/grafana/tempo/tempodb/encoding/common"
-	v0 "github.com/grafana/tempo/tempodb/encoding/v0"
-)
-
-func NewRecordReaderWriter() common.RecordReaderWriter {
-	return v0.NewRecordReaderWriter()
-}
