--- tempodb/encoding/v1test/fake/b51766bd-a395-448b-9684-7a4fdbd053aa/meta.json
@@ -1,1 +1,0 @@
-{"format":"v1","blockID":"b51766bd-a395-448b-9684-7a4fdbd053aa","minID":"B7UgV/dai66HiE94sd4C8A==","maxID":"/sKPSC9vuc93bLZsNeyUAQ==","tenantID":"fake","startTime":"2021-02-16T16:35:06.822173452-05:00","endTime":"2021-02-16T16:45:06.822173757-05:00","totalObjects":10,"size":760,"compactionLevel":0,"encoding":"zstd"}
