+++ tempodb/encoding/v2/object_test.go
@@ -1,4 +1,4 @@
-package v0
+package v2
 
 import (
 	"bytes"
